+++ HISTORY.rst
@@ -162,6 +162,8 @@
 - Add support for workflow transition field defaults.
 - Consolidate workflow transition field code into the transition model.
 - Add extra columns to show if a transition field has a default and a lookup.
+- Move the workflow instance detail link from the actions menu to the list
+  facet menu.
 
 4.6.3 (2024-03-28)
 ==================